development raid configuration
